# applications/services/api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for school project",
        "title": "Book \u0026 Venyl Loan Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/book/all/": {
            "get": {
                "description": "Gets a list of all books",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Book"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/checkout/{title}": {
            "get": {
                "description": "Checkouts a book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.BookTitle"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/get-recs-author/{title}": {
            "get": {
                "description": "Gets a recommended list of all books from author by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.BookSimple"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/get-recs-year/{title}": {
            "get": {
                "description": "Gets a recommended list of all books from year by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.BookSimple"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/get-simple/{title}": {
            "get": {
                "description": "Gets a simplified list of books by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.BookSimple"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/get/{title}": {
            "get": {
                "description": "Gets a book by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Book"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/return/{title}": {
            "get": {
                "description": "Returns a book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.BookTitle"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/search/{title}": {
            "get": {
                "description": "Gets a list of all searched books by title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of book",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Book"
                            }
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
        "/book/write-csv-to-db/": {
            "get": {
                "description": "Writes a csv file to the db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Book"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.BookTitle"
                        }
                    },
                    "404": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.Book": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "author": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isbn": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "main.BookSimple": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "isbn": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "main.BookTitle": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}